Do not write past buffer in USB_ReadPacket()
USB_ReadPacket() copies data in multiple of 4-bytes. This can lead to
writing past the destination buffer if len is not multiple of 4.

The issue was observed in MSC class driver. USBH_MSC_ClassRequest()
passes pointer to MSC_Handler->max_lun as control transfer buffer.
Because max_lun variable is 4 bytes in length, the issue is causing as
severe problems as it could. USB_ReadPacket() correctly wrote the least
significant byte of the max_len but also did write the upper 3 bytes with
unwanted non-zero data. As a result, USB Host did incorrectly set the
max_lun value to MAX_SUPPORTED_LUN (2) and queried non-existed LUNs from
devices with just a single LUN.

Solve the issue by writing the last data word in byte mode if buffer
length is not multiple of 4.